requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
wandb==0.17.4
numpy>1.24
pandas==2.2.2
tqdm==4.66.4
seaborn==0.13.2
qiskit==1.2.0
ml-collections==0.1.1
transformers==4.42.1
cuda-quantum==0.8.0
pylatexenc==2.10
ipywidgets==8.1.3
numba==0.60.0
pyarrow==16.1.0
lightning==2.3.3
wheel==0.43.0
ipykernel==6.29.5
torch_geometric==2.5.3
jupyterlab